+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- feat(plugin-api): add runtime variables to request tab context. (`3c72bf5`)
 - feat(settings): enhance settings UI with new icons and layout improvements. (`52a033b`)
 - feat(plugin-api): enhance request editor and plugin context. (`58a3176`)